pyproject.toml
[tool.poetry]
name = "douban-group"
version = "0.1.0"
description = ""
authors = ["popflag <18240463+popflag@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "4.11.1"
requests = "2.28.1"
openai = "^1.35.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
